When not building statistics, process files' data right after scan.
Instead of gathering all the files' data and then process them,
process them on the go as the files are scanned.

This allows to feed the list of files to scan via stdin gradually,
leaving cowstats to elaborate them as they are sent.

This doesn't work with statistics view yet.