Add an option to ignore profiling variables.
When building with profiling support enabled with GCC, there are some
gcov-specific variables that are added to all the object files, some
of which are in the .bss section.

To avoid reporting these, provide a -p switch to disable them, akin to
the -x switch for C++'s false positives.