Improve the scan of object files for CoW symbols.
Instead of just checking by name of the section in which the symbol is
defined, make use of type and flags of the sections to choose what to
do with the symbol.

Handle .bss (and .tbss) straight away, and then start filtering the
rest of sections that are not to be scanned.

Warn when a symbol is found on a section that has unknown name.
Comment the ignored cases.