Improve support for unknown section types.
Accept unknown section types in the specific ranges, creating a dummy
unknown section type for them (as they are valid per the ELF
specification).

Raise the correct exception in case the unknown section type is
outside the specific ranges, as it means that either the code has to
be updated or the file is not valid per ELF specifications.

Add a new test to ensure the proper raise of the exception; also make
sure it's not raised by unknown section types in the specific range
(User-specific range in this case).