Rework again the way exclusions work.
Create two arrays: one for excluded _names_ and one for excluded
_Regexps_, this way it's possible to use the include? method on the
array instead of doing the double iteration over it.

Instead of creating a Set of all the used symbols, just add them to
the same $exclude_names array, since the effect of the two is the
same.

Make the array uniq before starting the main iteration over the
symbols.

Regular expressions are matched exactly like before.