Implement a module elf/tool.rb that allows easier tools writing.
Since all the analysis tools that ship with Ruby-Elf use a very
similar syntax, by using a new module, it's possible to reduce the
code duplication, especially in error handling and other similar
issues.

This is the first step toward having a much easier handling of tools
in Ruby-Elf.

Please note that this changes the cowstats parameter interface since
the --filelist/-f option has been removed in favour of just prefixing
the list file name with the @ symbol, similarly to how it's done by
other tools like size(1).