Change handling of ELF files without string tables.
Since static executables produced with assemblers like nasm and yasm will
not emit a proper string table, and will set the section header string
table index to zero, try to handle those anyway.

Divide the exception of missing string table in two, one stating that the
load of the file is not complete yet (that will short circuit the library
in case of errors in the load logic) and another that will actually mean an
absolute lack of string table.

Also, change the behaviour of the has_section? request so that if the file
has no shstrtab it just answers negatively rather than raising an
exception. This changes the behaviour of the library (and thus the
testsuite) but it should not influence the way the tools work and instead
should reduce the useless noise due to the exception raised before.