Change the way the path is validated, raise Errno::EISDIR or Errno::EINVAL.
This way we're satisfied that a path is not a valid file, or even is a
directory rather than a file. This falls in line with other libraries as
well as the Errno::ENOENT idea.

Also add tests to make sure that we can handle those error cases. Note that
this requires us to create the fifo ourselves, as well as supporting
absolute files to be opened.