Avoid using Range#min as it's broken on Ruby 1.8.
With Ruby 1.8, when using big values such as the reserved range for
sections, the Range#min method will try to iterate over all the elements to
find the smallest one, which is obviously not smart.

It's fixed in Ruby 1.9, but for now work it around by using Range#first
that works fine.

http://blog.aizatto.com/2006/11/11/unexpected-behavior-of-rubys-rangemin-rangemax/