missingstatic: cleanup and improve general performance.
First of all, drop before/after options and simply use initialize to set
all the defaults.

Then, make sure each symbol is only present once in the list (to avoid
having too big a list of symbols to exclude).

Also compile all the given regexp patterns into a single regular expression
and, finally, try compiling the list of symbols in one single big regular
expression.

Note: this *will* cause problems if there are too many used symbols in the
list, but until that can be safely reproduced, this works fine for
reasonably-sized projects (such as FFmpeg) and with JRuby (that doesn't
seem to have Regexp size limits, or has a much higher one).