link-collisions: rewrite harvester using the Elf::Tool base class.
This reduces tremendously the duplicated code between Elf::Tool and the
harvester, but also the harvester now uses PL/SQL on PostgreSQL to reduce
the number of round-trips between Ruby and PostgreSQL when adding new
implementations and symbols.

Also this ensures that each object will add all its symbols with a single
transaction, committed before closing the object, so that no
halfway-through transactions are left when a file is not properly loaded.

Note: starting from this version, the harvester is _always_ recursive in
subdirectories.