rbelf-nm: sort and mux output
With this change, the identified symbols are buffered until the whole file
is processed before they are printed, allowing multiple threads to work at
the same time.

This also permitted implementing sorting of symbols by name and address,
changing the default to sort by name, as the other nm(1) implementations
do.

Please note that while this reduces the feature difference between rbelf-nm
and GNU nm, the parameters are not compatible, as rbelf-nm uses the same
--sort-by syntax that cowstats uses, rather than the -P/-n options used by
GNU nm.