elfgrep: implement a -f option, and make pattern detection saner
The -f option allows passing a list of expressions from a file, one per
line, either fixed or regular. Even though the documentation in the grep
man page does not make it clear, it can be used as many times as needed and
mixing it with --regexp (exactly like grep(1)).

Now, to decide whether to look for a pattern in the first argument or not,
we check if either --regexp or --file was passed, as the latter can be
used, without any match at all, and would still invalidate the use of the
first argument as pattern.